+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### New features
 
-* [#187](https://github.com/clojure-emacs/inf-clojure/pull/197): Defvar `inf-clojure-eldoc-enabledp` to disable eldoc interaction.
+* [#187](https://github.com/clojure-emacs/inf-clojure/pull/197): Defcustom `inf-clojure-enable-eldoc` to disable eldoc interaction.
 
 ## 3.1.0 (2021-07-23)